+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 1.1.3
 
-- Very Minor changes
+- Tests for Worker